pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-cli"
version = "0.0.1"
description = "CLI for AI agents to interact with Reddit"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "leo", email = "78882874+leog25@users.noreply.github.com"}]
keywords = ["reddit", "cli", "ai", "agents", "scraping"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0,<1.0.0",
    "httpx>=0.24.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "rich>=13.0",
    "browser-cookie3>=0.19",
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/leog25/reddit-cli"
Issues = "https://github.com/leog25/reddit-cli/issues"
Releases = "https://github.com/leog25/reddit-cli-releases"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.4",
]
binary = [
    "rookiepy>=0.5",
]

[project.scripts]
reddit = "reddit_cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]
